+++ core/src/main/java/apoc/util/DateParseUtil.java
@@ -18,14 +18,29 @@ import java.time.temporal.TemporalAccessor;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
+import static apoc.ApocConfig.apocConfig;
 import static apoc.util.Util.getFormat;
+import static org.neo4j.configuration.GraphDatabaseSettings.db_temporal_timezone;
 
 public class DateParseUtil {
+    public static final String DB_TEMPORAL_TIMEZONE = apocConfig().getString(db_temporal_timezone.name());
+    
     private static Map<Class<? extends TemporalAccessor>, MethodHandle> parseDateMap = new ConcurrentHashMap<>();
     private static Map<Class<? extends TemporalAccessor>, MethodHandle> simpleParseDateMap = new ConcurrentHashMap<>();
     private static String METHOD_NAME = "parse";
+
+    public static ZoneId getTimezoneIfValid(Map<String, Object> config, String defaultZone) {
+        try {
+            return Optional.ofNullable((String) config.getOrDefault("timezone", defaultZone))
+                    .map(ZoneId::of)
+                    .orElse(null);
+        } catch (DateTimeException e) {
+            throw new IllegalArgumentException(String.format("The timezone field contains an error: %s", e.getMessage()));
+        }
+    }
 
     public static TemporalAccessor dateParse(String value, Class<? extends TemporalAccessor> date, String...formats) {
         return dateParse(value, date, null, formats);
@@ -76,7 +91,6 @@ public class DateParseUtil {
                     return LocalDateTime.parse(value, format).atZone(zoneId);
                 }
                 if (date.equals(OffsetTime.class)) {
-                    // todo - check this
                     return LocalTime.parse(value, format).atOffset(zoneId.getRules().getOffset(Instant.now()));
                 }
             }
