+++ docs/asciidoc/modules/ROOT/partials/usage/config/apoc.load.csv.adoc
@@ -16,7 +16,7 @@ The procedure support the following config parameters:
 | mapping | Map | {} | per field mapping, entry key is field name, .e.g `{years:{....}` see below
 | compression | `Enum[NONE, BYTES, GZIP, BZIP2, DEFLATE, BLOCK_LZ4, FRAMED_SNAPPY]` | `null` | Allow taking binary data, either not compressed (value: `NONE`) or compressed (other values)
 See the xref::overview/apoc.load/apoc.load.csv.adoc#_binary_file[Binary file example]
-| timezone | String | null | default timezone in case of localtime and localdatetime without this one
+| timezone | String | `db.temporal.timezone` from `neo4j.conf` | default timezone in case of localtime and localdatetime without this one
 |===
 
 `mapping` supports the following values:
