+++ docs/asciidoc/modules/ROOT/partials/usage/config/apoc.load.json.adoc
@@ -6,7 +6,7 @@ The procedure support the following config parameters:
 | ignore | List<String> | [] | which field to ignore
 | nullValues | List<String> | [] | which values to treat as null, e.g. `['na',false]`
 | mapping | Map | {} | per field mapping, entry key is field name, .e.g `{years:{....}` see below
-| timezone | String | null | default timezone in case of localtime and localdatetime without this one
+| timezone | String | `db.temporal.timezone` from `neo4j.conf` | default timezone in case of localtime and localdatetime without this one
 |===
 
 `mapping` supports the following values:
